Bob_backend/requirements.txt
Flask-Cors==4.0.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
python-dotenv==1.0.1
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-auth==2.32.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.8
langchain-core==0.2.19
langchain-google-genai==1.0.7
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.2
langsmith==0.1.86
lz4==4.3.3
multidict==6.0.5
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pinecone-client==3.2.2
pinecone-plugin-interface==0.0.7
proto-plus==1.24.0
protobuf==4.21.6
protoc-gen-openapiv2==0.0.1
pydantic==2.8.2
pydantic-core==2.20.1
tenacity==8.5.0
tqdm==4.66.4
typing-extensions==4.12.2
yarl==1.9.4
